添加两台蒸汽大机器

+++ src/main/java/com/science/gtnl/common/block/Casings/Casing/MetaBlockConstructors.java
@@ -9,28 +9,28 @@ import gregtech.api.render.TextureFactory;
 
 public final class MetaBlockConstructors {
 
-    public static ItemStack initMetaBlock(String originEnglishName, byte meta, MetaCasingBase basicBlock) {
+    public static ItemStack initMetaBlock(String originEnglishName, int meta, MetaCasingBase basicBlock) {
         return MetaItemStackUtils.initMetaItemStack(originEnglishName, meta, basicBlock, basicBlock.getUsedMetaSet());
     }
 
-    public static ItemStack initMetaBlock(String originEnglishName, byte meta, MetaCasingBase basicBlock,
+    public static ItemStack initMetaBlock(String originEnglishName, int meta, MetaCasingBase basicBlock,
         String[] tooltips) {
         basicBlock.getTooltipsMap()
-            .put((int) meta, tooltips);
+            .put(meta, tooltips);
         return MetaItemStackUtils.initMetaItemStack(originEnglishName, meta, basicBlock, basicBlock.getUsedMetaSet());
     }
 
-    public static void setCasingTextureForMetaBlock(byte meta, MetaBlockCasingBase basicBlock) {
+    public static void setCasingTextureForMetaBlock(int meta, MetaBlockCasingBase basicBlock) {
         Textures.BlockIcons
             .setCasingTextureForId(basicBlock.getTextureIndex(meta), TextureFactory.of(basicBlock, meta));
     }
 
-    public static ItemStack initMetaBlockCasing(String originEnglishName, byte meta, MetaBlockCasingBase basicBlock) {
+    public static ItemStack initMetaBlockCasing(String originEnglishName, int meta, MetaBlockCasingBase basicBlock) {
         setCasingTextureForMetaBlock(meta, basicBlock);
         return initMetaBlock(originEnglishName, meta, basicBlock);
     }
 
-    public static ItemStack initMetaBlockCasing(String originEnglishName, byte meta, MetaBlockCasingBase basicBlock,
+    public static ItemStack initMetaBlockCasing(String originEnglishName, int meta, MetaBlockCasingBase basicBlock,
         String[] tooltips) {
         setCasingTextureForMetaBlock(meta, basicBlock);
         return initMetaBlock(originEnglishName, meta, basicBlock, tooltips);
